feat: import TSV file with merging based on composite key of STB/TITLE/DATE
Imports a TSV file and merges it into a denormalized TSV file that contains
the merge of all imported TSV files.  An entry from the import file replaces
an entry in the data store when the composite key STB|TITLE|DATE appears
in the datastore already.  The composite key is composed of trimmed values.

The data is stored in a single denormalized TSV file because this is the
simplest method of storing this data.  Space could be saved, and queries
could be made faster, by denormalizing the data.  However, it seems unlikely
that a TSV file would be used for a backend on a system like this so I opted
to do the simplest thing.

This commit also includes some test import data.