feat: implement query script, which allows selecting zero or more fields, zero or more fields to order by, and zero or 1 fields to filter on.
When zero fields are selected, output defaults to all.  When zero fields
are specified for ordering, no ordering is performed.  When zero fields are
specified for filtering, no filtering is performed.

The query is performed using CSV getline, which performs a line-by-line
read from the CSV file rather than loading it all into memory.